backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.0
openpyxl==3.1.2
textract==1.6.5
openai==0.28.0
